pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
replit = "^3.2.4"
Flask = "^2.1.3"
Flask-SocketIO = "^5.3.0"
python-dotenv = "^0.21.0"
pytz = "^2022.6"
SQLAlchemy = "^1.4.42"
faunadb = "^4.3.1"
Flask-Login = "^0.6.2"
Flask-Script = "^2.0.6"
flask-sqlalchemy = "^3.0.2"
hatesonar = "^0.0.7"

[tool.poetry.dev-dependencies]
python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.0"}
debugpy = "^1.6.2"
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"